refactor: split collapsible primitives
